Revert "merge dev into main (async call for product detail) (#1)"
This reverts commit 174ef081bedb79167339da61788270bccdbd0d47.
