+++ requirements.txt
@@ -1,14 +1,13 @@
 # Default
 selenium==4.33.0
 webdriver-manager==4.0.2
-pytest-playwright>=0.7.0
+pytest-playwright>=1.5.0
 beautifulsoup4==4.13.4
 pandas>=2.2.3
 
 # DB
 sqlalchemy>=2.0.41
 psycopg[binary]>=3.2.0
-asyncpg>=0.28.0
 
 # for testing
 behave==1.2.6
