+++ scraper/__init__.py
@@ -1,3 +1,1 @@
 # scraper/__init__.py
-
-from .selector import Selector,ProductDetailSelector
