+++ scraper/model.py
