+++ scraper/selector.py
@@ -16,5 +16,6 @@ class ProductDetailSelector:
     NAME: str = 'h1.fp-page-header.fp-page-title'
     PRICE: str = 'div.fp-item-detail.fp-item-detail-lg div.fp-item-price span.fp-item-base-price'
     UNIT: str = 'div.fp-item-price span.fp-item-size'
-    LOCATION: str = 'div.fp-item-location'
+    
+    LOCATIOIN: str = 'div.fp-item-location'
     SKU: str = 'div.fp-item-upc'
